Consolidation
Removes unnecessary folders
