--- src/main/Main.java
@@ -1,10 +1,0 @@
-package main;
-
-public class Main {
-
-  public static void main(String[] args) {
-
-    System.out.println("Video and Chill");
-
-  }
-}
